backend/requirements.txt
# Flask Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1

# Database - SQLite (built into Python)
SQLAlchemy==2.0.23
alembic==1.13.1

# Configuration & Environment
python-dotenv==1.0.0
pydantic==2.5.2

# Security
bcrypt==4.1.2
cryptography==41.0.8

# Testing
pytest==7.4.3
pytest-flask==1.3.0

# Development
black==23.11.0
flake8==6.1.0
isort==5.12.0

# ASI:One LLM Integration (OpenAI-compatible)
llama-index-llms-openai-like==0.4.0
llama-index-core==0.10.0

# Additional LLM utilities
pydantic==2.5.2
aiohttp==3.9.1
asyncio-throttle==1.0.2

# Fetch.ai uAgents Framework - Epic 2.2
uagents==0.22.5
cosmpy==0.9.2
msgpack==1.0.8 